Extracted ledger login into separate component
