+++ src/components/Login/LoginFormLedger/LoginFormLedger.scss
@@ -1,0 +1,11 @@
+.loginForm {
+  label {
+    display: block;
+  }
+
+  .actions {
+    margin-top: 24px;
+    padding-top: 16px;
+    border-top: 1px solid rgba(#000, 0.1);
+  }
+}
