+++ src/components/Login/index.js
@@ -1,9 +1,7 @@
-import { compose } from 'recompose';
-import { withData, withError, withActions } from 'spunky';
+import { withActions } from 'spunky';
 
 import Login from './Login';
 import authActions from '../../actions/authActions';
-import ledgerActions from '../../actions/ledgerActions';
 
 const mapAuthActionsToProps = (actions) => ({
   login: ({ wif, passphrase, encryptedWIF, publicKey }) => {
@@ -11,22 +9,4 @@ const mapAuthActionsToProps = (actions) => ({
   }
 });
 
-const mapLedgerActionsToProps = (actions) => ({
-  poll: () => actions.call()
-});
-
-const mapLedgerDataToProps = (data) => {
-  const { deviceInfo, publicKey } = data || {};
-  return { deviceInfo, publicKey };
-};
-
-const mapLedgerErrorToProps = (error) => ({
-  deviceError: error
-});
-
-export default compose(
-  withActions(authActions, mapAuthActionsToProps),
-  withActions(ledgerActions, mapLedgerActionsToProps),
-  withData(ledgerActions, mapLedgerDataToProps),
-  withError(ledgerActions, mapLedgerErrorToProps)
-)(Login);
+export default withActions(authActions, mapAuthActionsToProps)(Login);
